projet-chatbot/chatbot_app/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-dance
google-genai
gunicorn
psycopg2-binary
oauthlib
pyjwt
sift-stack-py
sqlalchemy
werkzeug
email-validator
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
requests
beautifulsoup4
Flask-Mail
Flask-Migrate